accounts test

+++ app/src/test/java/restapitester/ClientService.java
@@ -1,7 +1,6 @@
 package restapitester;
 
 
-import restapitester.api.Client;
 import retrofit2.Call;
 
 import retrofit2.http.*;
@@ -11,7 +10,7 @@ import java.util.List;
 public interface ClientService {
     @GET("client/{id}")
     @Headers("X-API-VERSION:1")
-    Call<Client> getClientById(@Path("id") Integer id);
+    Call<Client> getClientById(@Path("id") long id);
 
     @GET("client")
     @Headers("X-API-VERSION:1")
@@ -23,7 +22,7 @@ public interface ClientService {
 
     @DELETE("client/{id}")
     @Headers("X-API-VERSION:1")
-    Call deleteClientById(@Path("id") Integer id);
+    Call deleteClientById(@Path("id") long id);
 
     @DELETE("client/login/{clientLogin}")
     @Headers("X-API-VERSION:1")
